Updated broken endpoint URL
Hey there - noticed that the endpoint URL was broken. I work at Alchemy and in checking Astar's recommended endpoints (here - https://docs.astar.network/docs/quickstart/endpoints/), Alchemy is listed as the first one and the only one that is not a public endpoint. As public endpoints are often rate-limited and therefore not great for production environments, I added the Alchemy endpoint URL to help developers going through this doc. Lmk of any questions, thanks!