Fixing the contract lowercasing confusion
I erroneously assumed that lowercasing the contract addresses is
a requirement and edited docs with that in mind. The real issue
arises when same variable are used for subscribing and item
filtering, because in the archive output contract addresses are
always in lower case.

Fixing this misunderstanding by rewriting the warnings at
docs/evm-indexing/configuration and
docs/substrate-indexing/evm-support,
and also removing the superfluous .toLowerCase() calls.

	modified:   docs/evm-indexing/batch-processor-in-action.md
	modified:   docs/evm-indexing/configuration.md
	modified:   docs/evm-indexing/context-interfaces.md
	modified:   docs/evm-indexing/squid-evm-typegen.md
	modified:   docs/substrate-indexing/evm-support.md